Add FacialDirectionDataset
Added the FacialDirectionDataset, which consists of images of faces with corresponding labels indicating the direction each face is facing (left, right, up, straight). The dataset is in CSV format, with each row representing an image and its label. The images have been preprocessed to a size of 64x64 pixels and are grayscale. This dataset will be used for training and evaluating neural networks for facial direction classification tasks.